requirements.txt
# ─── Core ML ────────────────────────────────────────────────────────────────
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0
joblib>=1.3.0

# ─── API ────────────────────────────────────────────────────────────────────
flask>=3.0.0
gunicorn>=21.2.0

# ─── Data Validation ────────────────────────────────────────────────────────
pydantic>=2.5.0

# ─── Environment & Config ────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ─── Experiment Tracking (optional — install separately if OpenMP available)
# mlflow==2.9.2    # uncomment after: brew install libomp  (macOS)
#                  # or on Linux/Docker it installs automatically

# ─── Testing ────────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-flask>=1.2.0

# ─── HTTP Client (for health checks) ────────────────────────────────────────
requests>=2.31.0
